test: move code under scheduler_perf
Some code in test/utils/runners.go was only used by scheduler_perf.
Therefore it belongs under test/integration/scheduler_perf where it can be
maintained by SIG Scheduling.
